Simplified solution using Comlink
