+++ src/renderer/components/CrossBrowser/Screenshots/index.js
@@ -1,10 +1,7 @@
-import playwright from 'playwright'
 import { reactive, watchEffect } from '@vue/composition-api'
-import { v1 as uuid } from 'uuid'
-import playwrightWorker from '@/workers/index'
-
-const test = playwrightWorker(50)
-console.log(test)
+// Exposes the Web Worker using Comlink
+// https://github.com/GoogleChromeLabs/comlink-loader#singleton-mode
+import { CrossBrowserScreenshot } from '@/workers/playwright.worker'
 
 // Keep track of all the open browser contexts
 // This data can be accessed reactively
@@ -16,37 +13,117 @@ export const browserContexts = reactive({
 watchEffect(() => console.log('browser contexts', browserContexts.active))
 
 // A single browser's screenshot
-class CrossBrowserScreenshot {
-  constructor(options) {
-    this.url = options.url || 'https://google.com'
-    this.browser = ['chromium', 'webkit', 'firefox'].includes(options.browser)
-      ? options.browser
-      : 'chromium' // Default to Chromium
-    this.height = options.height || 100
-    this.width = options.width || 100
-    this.x = options.x || 0
-    this.y = options.y || 0
+// class CrossBrowserScreenshot {
+//   constructor(options) {
+//     this.url = options.url || 'https://google.com'
+//     this.browser = ['chromium', 'webkit', 'firefox'].includes(options.browser)
+//       ? options.browser
+//       : 'chromium' // Default to Chromium
+//     this.height = options.height || 100
+//     this.width = options.width || 100
+//     this.x = options.x || 0
+//     this.y = options.y || 0
 
-    this.isLoading = true // Initial state
-  }
+//     this.isLoading = true // Initial state
+//   }
 
-  async takeScreenshot() {
-    const id = uuid()
-    console.log(`Preparing to launch ${this.browser}, ${id}`)
-    this.isLoading = true // Update loading state
+//   async takeScreenshot() {
+//     const id = uuid()
+//     console.log(`Preparing to launch ${this.browser}, ${id}`)
+//     this.isLoading = true // Update loading state
 
-    const browser = await playwright[this.browser].launch({
-      /* headless: false */
+//     const browser = await playwright[this.browser].launch({
+//       /* headless: false */
+//     })
+//     const context = await browser.newContext({
+//       viewport: {
+//         height: this.height,
+//         width: this.width,
+//       },
+//     })
+
+//     // Track the context
+//     browserContexts.active.push({ id, context, type: this.browser })
+
+//     // Remove this ID from the list of active
+//     function removeBrowserContext(id) {
+//       browserContexts.active = browserContexts.active.filter(
+//         (i) => !i.id === id
+//       )
+//     }
+
+//     try {
+//       console.log(`Loading ${this.browser}, ${id}`)
+
+//       const page = await context.newPage()
+//       await page.goto(this.url, {
+//         // waitUntil: 'networkidle',
+//       })
+
+//       console.log('x,y', [this.x, this.y])
+
+//       // Scroll up/down as needed
+//       await page.evaluate(`window.scrollTo(${this.x}, ${this.y})`)
+
+//       // Wait...
+//       await page.waitForTimeout(1000)
+
+//       const screenshotBuffer = await page.screenshot({
+//         //   clip: {
+//         //     x: 0,
+//         //     y: 0,
+//         //     height: this.height,
+//         //     width: this.width,
+//         //   },
+//       })
+
+//       removeBrowserContext(id) // Remove from the array of active browsers
+//       await browser.close()
+
+//       return screenshotBuffer
+//     } catch (err) {
+//       this.isLoading = false
+//       removeBrowserContext(id)
+//       await browser.close()
+//       console.log(err)
+//       return false
+//     }
+//   }
+// }
+
+export async function takeScreenshots(
+  { url = '', browsers = [], height = 0, width = 0, x = 0, y = 0 },
+  callback = () => {}
+) {
+  // Loop through each browser
+  const screenshotPromiseGenerator = async (browserName) => {
+    // NOTE: The following NEEDS to have the await beforehand
+    // It is communicating with a Web Worker via Comlink
+    const instance = await new CrossBrowserScreenshot({
+      url,
+      browser: browserName,
+      height,
+      width,
+      x,
+      y,
     })
-    const context = await browser.newContext({
-      viewport: {
-        height: this.height,
-        width: this.width,
-      },
-    })
+    const { contextId } = instance
 
-    // Track the context
-    browserContexts.active.push({ id, context, type: this.browser })
+    try {
+      // Track the context
+      browserContexts.active.push({ id: contextId, type: browserName })
+
+      // Take the screenshot
+      const screenshots = await instance.takeScreenshot()
+
+      // Stop tracking this context
+      removeBrowserContext(contextId)
+
+      return { type: browserName, img: screenshots }
+    } catch (err) {
+      removeBrowserContext(contextId)
+      console.error(err)
+    }
 
     // Remove this ID from the list of active
     function removeBrowserContext(id) {
@@ -54,66 +131,6 @@ class CrossBrowserScreenshot {
         (i) => !i.id === id
       )
     }
-
-    try {
-      console.log(`Loading ${this.browser}, ${id}`)
-
-      const page = await context.newPage()
-      await page.goto(this.url, {
-        // waitUntil: 'networkidle',
-      })
-
-      console.log('x,y', [this.x, this.y])
-
-      // Scroll up/down as needed
-      await page.evaluate(`window.scrollTo(${this.x}, ${this.y})`)
-
-      // Wait...
-      await page.waitForTimeout(1000)
-
-      const screenshotBuffer = await page.screenshot({
-        //   clip: {
-        //     x: 0,
-        //     y: 0,
-        //     height: this.height,
-        //     width: this.width,
-        //   },
-      })
-
-      removeBrowserContext(id) // Remove from the array of active browsers
-      await browser.close()
-
-      return screenshotBuffer
-    } catch (err) {
-      this.isLoading = false
-      removeBrowserContext(id)
-      await browser.close()
-      console.log(err)
-      return false
-    }
-  }
-}
-
-export async function takeScreenshots(
-  { url = '', browsers = [], height = 0, width = 0, x = 0, y = 0 },
-  callback = () => {}
-) {
-  // Loop through each browser
-  const screenshotPromiseGenerator = (browser) => {
-    return new Promise((resolve) => {
-      const b = new CrossBrowserScreenshot({
-        url,
-        browser,
-        height,
-        width,
-        x,
-        y,
-      })
-
-      b.takeScreenshot().then((screenshot) => {
-        resolve({ type: browser, img: screenshot })
-      })
-    })
   }
 
   // Loop through each browser
