--- src/renderer/workers/index.js
@@ -1,14 +1,0 @@
-import PromiseWorker from 'promise-worker'
-// import Worker from 'worker-loader!./worker'
-import Worker from 'worker-loader!./worker'
-
-const worker = new Worker()
-const promiseWorker = new PromiseWorker(worker)
-
-const getPrimes = (amount) =>
-  promiseWorker.postMessage({
-    type: 'getPrimesMessage',
-    amount,
-  })
-
-export default getPrimes
